requirements.txt
sentence-transformers
faiss-cpu
streamlit
PyYAML
